i2c_rtc: avoid manually computing constants
The constants were manually computed to see if the following error
would go off:

    $ avr-gcc -Wall -Wpedantic -Wextra -O1 -mmcu=atmega32 -o rtc.out rtc.c
    i2c/i2c.c ../lcd_display/lcd/lcd.c
    In file included from i2c/i2c.c:2:0:
    /usr/lib/avr/include/util/delay.h: In function ‘_delay_us’:
    /usr/lib/avr/include/util/delay.h:245:2: error:
    __builtin_avr_delay_cycles expects a compile time integer constant
      __builtin_avr_delay_cycles(__ticks_dc);

It doesn't seem to be going off unless and until compiler optimizations
are turned off (which is weird!). So, there's no use in manually.
That' because computing the constant values as computing the values
programmatically with optimizations turned off results in the code
compiling correctly.

So, ...